refactor: HandStatus 클래스를 dto 패키지로 이동
