+++ src/test/java/domain/BlackJackResultTest.java
@@ -1,6 +1,7 @@
 package domain;
 
 import domain.constant.GamerResult;
+import domain.dto.HandStatus;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
